color estimated

+++ internal/brewbeer/processes/fermentation/fermentation_params.go
@@ -1,6 +1,6 @@
 package fermentation
 
 type Params struct {
-	InitialDensity float32
-	FinalDensity   float32
+	InitialDensity float64
+	FinalDensity   float64
 }
